refactor: split lib.rs into error, types, value, proc, info modules
Move the crate-root god module (~4.5k lines) into focused modules:
- error.rs — PmixError, PmixStatus
- types.rs — proc/job/data enums, IOFChannelFlags, BuilderError
- value.rs — payload/builder/owned value
- proc.rs — Proc handle
- info.rs — Info/InfoBuilder/InfoFlags + helpers

lib.rs keeps Context, init/put/get/fence/finalize, re-exports, and tests.
Crate root public API preserved via pub use.

Fixes #4.
